mina::to_ascii_decimal: to_ascii_decimal_f (25)
Sufficiently finished.

+++ mina/mina/detail/to_ascii_decimal_u.hpp
@@ -156,7 +156,9 @@ struct to_ascii_decimal_u<uint64_t> {
 		for (uint64_t d: xy) {
 			uint128_t acc(d);
 			acc += c;
-			c = (acc * divider_10e8) >> (64 + shift_10e8);
+			c = uint64_t(
+				(acc * divider_10e8) >> (64 + shift_10e8)
+			);
 			bv[2 - dp] = to_ascii_decimal_u<uint32_t>::to_bcd(
 				uint32_t(acc - c * 100000000)
 			)[1];
